vexpress: use ${CROSS_COMPILE} if set
For the vexpress platform, it is now possible to set CROSS_COMPILE
rather than CC_DIR. This makes it easier to call the script from
Makefile that already uses CROSS_COMPILE. The path to the runtime
libraries (libc and so on) is retrieved from the compiler itself, which
should be more robust. With the previous CC_DIR scheme, it was not
possible to build using the standard Ubuntu GCC package, for instance.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>