hikey: inittab: spawn login shell only on ${CFG_NW_CONSOLE_UART}
Currently, etc/inittab-hikey tries to spawn a login shell on both
ttyAMA0 and ttyAMA3 which doesn't work well when secure world and
normal world are configured to use different UARTs (in this case,
secure world won't log any more output as soon as Linux has booted).
This patch generates an inittab file with a single login shell  entry,
based on the value of $CFG_NW_CONSOLE_UART. If not set, ttyAMA3 is
used which is UART1 on the low-speed expension connector.

Fixes: https://github.com/OP-TEE/optee_os/issues/1483
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Joakim Bech <joakim.bech@linaro.org>